create proper directory structure
now that we're adding more functionality beyond basic Booth Finder, we need a proper structure
