+++ app/scrape.py
@@ -14,7 +14,7 @@ from webdriver_manager.chrome import ChromeDriverManager
 from webdriver_manager.core.utils import ChromeType
 from webdriver_manager.core.utils import read_version_from_cmd, PATTERN
 
-from hours import hour1, hour2
+from app.hours import hour1, hour2
 
 
 def scrape():
