+++ app/static/misc.js
